+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/Slots/DiffusionProfileInputMaterialSlot.cs
@@ -7,6 +7,7 @@ using UnityEngine;
 using UnityEngine.UIElements;
 using UnityEditor.ShaderGraph.Drawing.Controls;
 using UnityEngine.Experimental.Rendering.HDPipeline;
+using UnityEngine.Rendering;
 
 namespace UnityEditor.Experimental.Rendering.HDPipeline
 {
@@ -14,19 +15,21 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
     [FormerName("UnityEditor.ShaderGraph.DiffusionProfileInputMaterialSlot")]
     class DiffusionProfileInputMaterialSlot : Vector1MaterialSlot
     {
-        [SerializeField]
-        PopupList m_DiffusionProfile = new PopupList();
+        [SerializeField, Obsolete("Use m_DiffusionProfileAsset instead.")]
+        PopupList m_DiffusionProfile;
 
-        public PopupList diffusionProfile
+        [SerializeField]
+        DiffusionProfileSettings m_DiffusionProfileAsset;
+
+        [SerializeField]
+        int m_Version;
+        
+        DiffusionProfileSlotControlView view;
+
+        public DiffusionProfileSettings diffusionProfile
         {
-            get
-            {
-                return m_DiffusionProfile;
-            }
-            set
-            {
-                m_DiffusionProfile = value;
-            }
+            get { return m_DiffusionProfileAsset; }
+            set { m_DiffusionProfileAsset = value; }
         }
 
         public DiffusionProfileInputMaterialSlot()
@@ -37,46 +40,80 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                                           ShaderStageCapability stageCapability = ShaderStageCapability.All, bool hidden = false)
             : base(slotId, displayName, shaderOutputName, SlotType.Input, 0.0f, stageCapability, hidden: hidden)
         {
-            var hdPipeline = UnityEngine.Rendering.RenderPipelineManager.currentPipeline as HDRenderPipeline;
-            if (hdPipeline != null)
-            {
-                var diffusionProfileSettings = hdPipeline.diffusionProfileSettings;
-                m_DiffusionProfile.popupEntries = new List<string>();
-                m_DiffusionProfile.popupEntries.Add("None");
-
-                if (!hdPipeline.IsInternalDiffusionProfile(diffusionProfileSettings))
-                {
-                    var profiles = diffusionProfileSettings.profiles;
-                    for (int i = 0; i < profiles.Length; i++)
-                    {
-                        m_DiffusionProfile.popupEntries.Add(profiles[i].name);
-                    }
-                    m_DiffusionProfile.selectedEntry = Mathf.Min(m_DiffusionProfile.selectedEntry, profiles.Length + 1);
-                }
-                else
-                {
-                    m_DiffusionProfile.selectedEntry = 0;
-                }
-            }
         }
 
         public override VisualElement InstantiateControl()
         {
-            return new DiffusionProfileSlotControlView(this);
+            view = new DiffusionProfileSlotControlView(this);
+            return view;
+        }
+
+        AbstractMaterialNode matOwner;
+        public override void AddDefaultProperty(PropertyCollector properties, GenerationMode generationMode)
+        {
+            matOwner = owner as AbstractMaterialNode;
+            if (matOwner == null)
+                throw new Exception(string.Format("Slot {0} either has no owner, or the owner is not a {1}", this, typeof(AbstractMaterialNode)));
+
+            // Note: Unity ShaderLab can't parse float values with exponential notation so we just can't
+            // store the hash nor the asset GUID here :(
+            var diffusionProfileHash = new Vector1ShaderProperty
+            {
+                overrideReferenceName = "_DiffusionProfileHash",
+                hidden = true,
+                value = 0
+            };
+            var diffusionProfileAsset = new Vector4ShaderProperty
+            {
+                overrideReferenceName = "_DiffusionProfileAsset",
+                hidden = true,
+                value = Vector4.zero
+            };
+
+            properties.AddShaderProperty(diffusionProfileHash);
+            properties.AddShaderProperty(diffusionProfileAsset);
+
+            // We can't upgrade here because we need to access the current render pipeline asset which is not
+            // possible outside of unity context so we wait the next editor frame to do it
+            EditorApplication.update += UpgradeIfNeeded;
         }
 
         public override string GetDefaultValue(GenerationMode generationMode)
         {
-            return m_DiffusionProfile.selectedEntry.ToString();
+            if (m_DiffusionProfileAsset == null)
+                return "_DiffusionProfileHash";
+            else
+                return "((asuint(_DiffusionProfileHash) != 0) ? _DiffusionProfileHash : asfloat(uint(" + m_DiffusionProfileAsset.profile.hash + ")))";
         }
 
         public override void CopyValuesFrom(MaterialSlot foundSlot)
         {
             var slot = foundSlot as DiffusionProfileInputMaterialSlot;
+
             if (slot != null)
             {
-                diffusionProfile = slot.diffusionProfile;
+                m_DiffusionProfileAsset = slot.m_DiffusionProfileAsset;
             }
+        }
+
+        void UpgradeIfNeeded()
+        {
+#pragma warning disable 618
+            // Once the profile is upgraded, we set the selected entry to 0 (which was previously none
+            // in the diffusion profile index so it's fine if we don't upgrade it
+            if (m_Version == 0)
+            {
+                // We need to warn the user that we can't upgrade the diffusion profile but this upgrade code
+                // does not work currently :(
+                // Debug.LogError("Failed to upgrade the diffusion profile slot value, reseting to default value: " + hdAsset.diffusionProfileSettingsList[m_DiffusionProfile.selectedEntry] + 
+                //     "\nTo remove this message save the shader graph with the new diffusion profile reference");
+                // m_DiffusionProfileAsset = hdAsset.diffusionProfileSettingsList[m_DiffusionProfile.selectedEntry];
+                m_Version = 1;
+                // Sometimes the view is created after we upgrade the slot so we need to update it's value
+                view?.UpdateSlotValue();
+            }
+#pragma warning restore 618
+            EditorApplication.update -= UpgradeIfNeeded;
         }
     }
 }